requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
requests
google-generativeai
